Fix compilation on macOS using LLVM/Clang
Clang complains about not being able to initialize a variable of type
'const crn_format' with an rvalue of type 'int' in decrunch.cpp
generated from decrunch.pyx. This was probably due to the fact that the
according type crn_format was declared as int instead of enum in the
Cython source definition. This error was fixed by declaring this type
as an enum with its values taken from the C file.

Another error was the missing header file 'malloc.h' which is
considered deprecated (see [1]). 'stdlib.h' was used instead as it
should handle the platform specific includes of malloc.

Lastly malloc_usable_size was also missing on macOS, therefore it was
redeclared to malloc_size as multiple sources suggest (see [2] and [3]).

[1]: http://stackoverflow.com/a/12973361/1532986
[2]: https://bugzilla.mozilla.org/show_bug.cgi?id=419131#c0
[3]: http://opensnc.sourceforge.net/forum/viewtopic.php?pid=8603#p8603